requirements.txt
# requirements.txt

# Core application
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Data science stack
# Let scikit-learn determine the correct numpy and pandas versions
scikit-learn==1.7.2
pandas==2.2.2

# Supabase and Authentication
supabase==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.1
requests==2.31.0
